Move Kubernetes types to public module
